+++ src/domain/services/AuthConfirmService.ts
@@ -4,7 +4,6 @@ import {CrudService} from '@steroidsjs/nest/usecases/services/CrudService';
 import SearchQuery from '@steroidsjs/nest/usecases/base/SearchQuery';
 import {DataMapper} from '@steroidsjs/nest/usecases/helpers/DataMapper';
 import {ValidationHelper} from '@steroidsjs/nest/usecases/helpers/ValidationHelper';
-import {IUserService} from '@steroidsjs/nest-modules/user/services/IUserService';
 import {ModuleHelper} from '@steroidsjs/nest/infrastructure/helpers/ModuleHelper';
 import {AuthModule} from '@steroidsjs/nest-modules/auth/AuthModule';
 import {ContextDto} from '@steroidsjs/nest/usecases/dtos/ContextDto';
@@ -18,7 +17,8 @@ import {AUTH_CONFIRM_PROVIDERS_TOKEN, IAuthConfirmProvider} from '../interfaces/
 import {AuthConfirmSendCodeDto} from '../dtos/AuthConfirmSendCodeDto';
 import {AuthConfirmSaveDto} from '../dtos/AuthConfirmSaveDto';
 import {AuthConfirmProviderType} from '../types/AuthConfirmProviderType';
-import {GET_AUTH_CONFIRM_TARGET_FIELD_USE_CASE_TOKEN, IGetAuthConfirmTargetFieldUseCase} from '../../usecases/getAuthConfirmTargetField/IGetAuthConfirmTargetFieldUseCase';
+import {AuthConfirmTargetService} from "./AuthConfirmTargetService";
+
 
 @Injectable()
 export class AuthConfirmService extends CrudService<
@@ -33,10 +33,7 @@ export class AuthConfirmService extends CrudService<
         public repository: IAuthConfirmRepository,
         @Inject(AUTH_CONFIRM_PROVIDERS_TOKEN)
         protected readonly authConfirmProviders: IAuthConfirmProvider[],
-        @Inject(IUserService)
-        protected readonly userService: IUserService,
-        @Inject(GET_AUTH_CONFIRM_TARGET_FIELD_USE_CASE_TOKEN)
-        protected readonly getAuthConfirmTargetFieldUseCase: IGetAuthConfirmTargetFieldUseCase,
+        protected readonly authConfirmTargetService: AuthConfirmTargetService,
     ) {
         super();
     }
@@ -46,15 +43,19 @@ export class AuthConfirmService extends CrudService<
         providerType: AuthConfirmProviderType | null,
         context: ContextDto,
         schemaClass = null,
+        resolvedTargetInfo = null,
     ): Promise<AuthConfirmModel> {
         await ValidationHelper.validate(dto, {context});
+
         const config: IAuthConfirmConfig = ModuleHelper.getConfig<IAuthModuleConfig>(AuthModule).confirm;
 
         if (!providerType) {
             providerType = config.providerType;
         }
 
-        const targetField = this.getAuthConfirmTargetFieldUseCase.handle(providerType);
+        if (!resolvedTargetInfo) {
+            resolvedTargetInfo = await this.authConfirmTargetService.resolveTargetInfo(providerType, dto.target);
+        }
 
         // Не отправляем повторно смс, если она была отправлена недавно. Используем ту же модель
         // TODO Не уверен насколько это правильная логика.. Нужно подумать.
@@ -68,8 +69,9 @@ export class AuthConfirmService extends CrudService<
                         formatISO9075(addSeconds(new Date(), -1 * config.repeatLimitSec)),
                     ])
                     .andWhere({
-                        [targetField]: dto.target,
+                        target: resolvedTargetInfo.target,
                         isConfirmed: false,
+                        providerName: providerType,
                     }),
             );
             if (model) {
@@ -81,12 +83,14 @@ export class AuthConfirmService extends CrudService<
         if (config.isEnableDebugStaticCode) {
             code = _repeat('1', config.codeLength);
         } else {
-            const authConfirmProvider = this.authConfirmProviders.find(provider => provider.type === providerType);
+            const authConfirmProvider = this.authConfirmProviders
+                .find(provider => provider.type === providerType);
+
             if (!authConfirmProvider) {
                 throw new Error('Wrong provider type: ' + providerType);
             }
 
-            code = await authConfirmProvider.generateAndSendCode(config, dto.target);
+            code = await authConfirmProvider.generateAndSendCode(config, resolvedTargetInfo.target);
 
             if (!code) {
                 throw new Error('Code is not sent, provider type: ' + providerType);
@@ -96,7 +100,7 @@ export class AuthConfirmService extends CrudService<
         // Сохраняем в БД
         const model = await this.repository.create(
             DataMapper.create(AuthConfirmModel, {
-                [targetField]: dto.target,
+                target: resolvedTargetInfo.target,
                 code,
                 providerName: providerType,
                 expireTime: formatISO9075(addMinutes(new Date(), config.expireMins)),